Jack update

+++ src/main/java/com/paipianwang/pat/facade/information/service/impl/PmsNewsServiceImpl.java
@@ -1,7 +1,6 @@
 package com.paipianwang.pat.facade.information.service.impl;
 
 
-import java.util.List;
 import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,6 +8,7 @@ import org.springframework.stereotype.Service;
 
 import com.paipianwang.pat.common.entity.DataGrid;
 import com.paipianwang.pat.common.entity.PageParam;
+import com.paipianwang.pat.common.util.ValidateUtil;
 import com.paipianwang.pat.facade.information.entity.PmsNews;
 import com.paipianwang.pat.facade.information.service.PmsNewsFacade;
 import com.paipianwang.pat.facade.information.service.biz.PmsNewsBiz;
@@ -22,12 +22,6 @@ public class PmsNewsServiceImpl implements PmsNewsFacade {
 
 	@Autowired
 	private final PmsNewsBiz biz = null;
-
-	@Override
-	public List<PmsNews> RecommendNews() {
-		final List<PmsNews> list = biz.RecommendNews(); 
-		return list;
-	}
 
 	@Override
 	public DataGrid<PmsNews> listWithPagination(PageParam pageParam, Map<String, Object> paramMap) {
@@ -49,13 +43,28 @@ public class PmsNewsServiceImpl implements PmsNewsFacade {
 
 	@Override
 	public long deleteByIds(long[] ids) {
-		final long ret = biz.deleteByIds(ids);
+		long ret = 0l;
+		if(ValidateUtil.isValid(ids)) {
+			ret = biz.deleteByIds(ids);
+		}
 		return ret;
 	}
 
 	@Override
 	public PmsNews findNewsById(long newsId) {
 		return biz.findNewsById(newsId);
+	}
+
+	@Override
+	public PmsNews findNextNew(final String tags, final int newId, final Integer recommend) {
+		final PmsNews news = biz.findNextNew(tags, newId, recommend);
+		return news;
+	}
+
+	@Override
+	public PmsNews findPreNew(final String tags, final int newId, final Integer recommend) {
+		final PmsNews news = biz.findPreNew(tags, newId, recommend);
+		return news;
 	}
 
 }
